refactor: Http 헤더 정보 클래스 패키지 이동.
