+++ tomcat/src/main/java/org/apache/http/info/HttpHeaderName.java
@@ -1,4 +1,4 @@
-package org.apache.http;
+package org.apache.http.info;
 
 import java.util.Arrays;
 import java.util.Objects;
